Some refactoring of gradle resolver
