+++ src/scala/com/github/johnynek/bazel_deps/DepsModel.scala
@@ -1602,203 +1602,6 @@ object ResolverType {
     }
 }
 
-object TryMerge {
-  def tryMerge[T: TryMerge](debugName: Option[String], a: T, b: T): Try[T] = {
-    implicitly[TryMerge[T]].tryMerge(debugName, a, b)
-  }
-
-  implicit def tryOptMerge[T: TryMerge]: TryMerge[Option[T]] =
-    new TryMerge[Option[T]] {
-      def tryMerge(debugName: Option[String], left: Option[T], right: Option[T]): Try[Option[T]] = {
-        (left, right) match {
-          case (None, None)       => Success(None)
-          case (Some(l), Some(r)) => TryMerge.tryMerge(debugName, l, r).map(Some(_))
-          case (Some(l), None)    => Success(Some(l))
-          case (None, Some(r))    => Success(Some(r))
-        }
-      }
-    }
-
-  implicit def tryStringMapMerge[T: TryMerge]: TryMerge[Map[String, T]] =
-    new TryMerge[Map[String, T]] {
-      def tryMerge(
-        debugName: Option[String],
-          left: Map[String, T],
-          right: Map[String, T]
-      ): Try[Map[String, T]] = {
-        (left.keySet ++ right.keySet).foldLeft(Try(Map[String, T]())) {
-          case (prevM, nextK) =>
-            prevM.flatMap { m =>
-              val r: Try[T] = (left.get(nextK), right.get(nextK)) match {
-                case (None, None) =>
-                  Failure(new Exception("Shouldn't happen, key was in keyset"))
-                case (Some(l), Some(r)) => TryMerge.tryMerge(Some(debugName.map{p => s"$p:$nextK"}.getOrElse(nextK)), l, r)
-                case (Some(l), None)    => Success(l)
-                case (None, Some(r))    => Success(r)
-              }
-              r.map { innerV =>
-                m + ((nextK, innerV))
-              }
-            }
-        }
-      }
-    }
-}
-sealed trait TryMerge[T] {
-  def tryMerge(debugName: Option[String], left: T, right: T): Try[T]
-}
-
-
-object GradleLockDependency {
-  sealed trait VersionState
-  case object EqualVersionSpecified extends VersionState
-  case object LeftVersion extends VersionState
-  case object RightVersion extends VersionState
-
-  def resolveVersions(dependencyName: String)(
-      left: Option[String],
-      right: Option[String]
-  ): Try[VersionState] = {
-    (left, right) match {
-      case (None, None)                   => Success(EqualVersionSpecified)
-      case (Some(l), None)                => Success(LeftVersion)
-      case (None, Some(r))                => Success(RightVersion)
-      case (Some(l), Some(r)) if (l == r) => Success(EqualVersionSpecified)
-      case (Some(l), Some(r)) => {
-        println(
-          s"This should probably not be allowed... but we are going to pick a version conflict highest if we can for $dependencyName between $l, $r"
-        )
-        VersionConflictPolicy.Highest.resolve(
-          None,
-          Set(Version(l), Version(r))
-        ) match {
-          case Validated.Valid(v) =>
-            if (v.asString == l) {
-              Success(LeftVersion)
-            } else {
-              Success(RightVersion)
-            }
-          case Validated.Invalid(iv) =>
-            Failure(new Exception(s"Unable ot combine versions, $iv"))
-        }
-      }
-    }
-  }
-
-  implicit val mergeInst = new TryMerge[GradleLockDependency] {
-    def tryMerge(
-      debugName: Option[String],
-        left: GradleLockDependency,
-        right: GradleLockDependency
-    ): Try[GradleLockDependency] = {
-      lazy val mergedDependencies = Some(
-              (left.transitive.getOrElse(Nil) ++ right.transitive.getOrElse(
-                Nil
-              )).sorted.distinct
-            ).filter(_.nonEmpty)
-
-      for {
-        v <- resolveVersions(debugName.getOrElse("Unknown"))(left.locked, right.locked)
-        _ <-
-          if (left.project == right.project) Success(())
-          else
-            Failure(
-              new Exception(
-                s"Unable to merge due to incompatible project setting, had $left, $right"
-              )
-            )
-      } yield {
-        v match {
-          case EqualVersionSpecified =>
-          GradleLockDependency(
-            locked = left.locked,
-            project = left.project,
-            transitive = mergedDependencies
-          )
-          case LeftVersion => GradleLockDependency(
-            locked = left.locked,
-            project = left.project,
-            transitive = mergedDependencies
-          )
-          case RightVersion => GradleLockDependency(
-            locked = right.locked,
-            project = right.project,
-            transitive = mergedDependencies
-          )
-        }
-
-      }
-    }
-  }
-}
-
-case class GradleLockDependency(
-    locked: Option[String],
-    project: Option[Boolean],
-    transitive: Option[List[String]]
-)
-
-object GradleLockFile {
-  implicit val mergeInst = new TryMerge[GradleLockFile] {
-    def tryMerge(
-        debugName: Option[String],
-        left: GradleLockFile,
-        right: GradleLockFile
-    ): Try[GradleLockFile] = {
-      for {
-        annotationProcessor <- TryMerge.tryMerge(
-          debugName,
-          left.annotationProcessor,
-          right.annotationProcessor
-        )
-        compileClasspath <- TryMerge.tryMerge(
-          debugName,
-          left.compileClasspath,
-          right.compileClasspath
-        )
-        resolutionRules <- TryMerge.tryMerge(
-          debugName,
-          left.resolutionRules,
-          right.resolutionRules
-        )
-        runtimeClasspath <- TryMerge.tryMerge(
-          debugName,
-          left.runtimeClasspath,
-          right.runtimeClasspath
-        )
-        testCompileClasspath <- TryMerge.tryMerge(
-          debugName,
-          left.testCompileClasspath,
-          right.testCompileClasspath
-        )
-        testRuntimeClasspath <- TryMerge.tryMerge(
-          debugName,
-          left.testRuntimeClasspath,
-          right.testRuntimeClasspath
-        )
-      } yield {
-        GradleLockFile(
-          annotationProcessor,
-          compileClasspath,
-          resolutionRules,
-          runtimeClasspath,
-          testCompileClasspath,
-          testRuntimeClasspath
-        )
-      }
-    }
-  }
-  def empty = GradleLockFile(None, None, None, None, None, None)
-}
-case class GradleLockFile(
-    annotationProcessor: Option[Map[String, GradleLockDependency]],
-    compileClasspath: Option[Map[String, GradleLockDependency]],
-    resolutionRules: Option[Map[String, GradleLockDependency]],
-    runtimeClasspath: Option[Map[String, GradleLockDependency]],
-    testCompileClasspath: Option[Map[String, GradleLockDependency]],
-    testRuntimeClasspath: Option[Map[String, GradleLockDependency]]
-)
-
 case class Options(
     versionConflictPolicy: Option[VersionConflictPolicy],
     languages: Option[Set[Language]],
