+++ src/scala/com/github/johnynek/bazel_deps/GradleResolver.scala
@@ -1,6 +1,7 @@
 package com.github.johnynek.bazel_deps
 
 import cats.MonadError
+import cats.data.{Validated, ValidatedNel}
 import coursier.util.Task
 import io.circe.jawn.JawnParser
 import cats.implicits._
@@ -16,11 +17,15 @@ import cats.implicits._
 
 class GradleResolver(
     servers: List[DependencyServer],
+    versionConflictPolicy: VersionConflictPolicy,
     ec: ExecutionContext,
     runTimeout: Duration,
     gradleTpe: ResolverType.Gradle,
     cachePath: Path
 ) extends Resolver[Task] {
+
+  implicit val tryMergeGradleLock = GradleLockDependency.mergeGradleLockDeps(versionConflictPolicy)
+
   private[this] lazy val coursierResolver =
     new CoursierResolver(servers, ec, runTimeout, cachePath)
 
@@ -32,29 +37,27 @@ class GradleResolver(
   ): Task[SortedMap[MavenCoordinate, ResolvedShasValue]] =
     coursierResolver.getShas(m)
 
-  private def loadLockFiles(
-      lockFiles: List[String]
-  ): Try[List[GradleLockFile]] =
-    lockFiles.traverse { next =>
-      (Model.readFile(new File(next)) match {
-        case Success(str) => Success(str)
-        case Failure(err) =>
-          Failure(new Exception(s"Failed to read ${next}", err))
-      })
-      .flatMap { content =>
-        Decoders.decodeGradleLockFile(new JawnParser, content) match {
-          case Right(m) => Success(m)
-          case Left(err) =>
-            Failure(new Exception(s"Failed to parse ${next}", err))
-        }
+  private def loadLockFile(
+      lockFile: String
+  ): Try[GradleLockFile] =
+    (Model.readFile(new File(lockFile)) match {
+      case Success(str) => Success(str)
+      case Failure(err) =>
+        Failure(new Exception(s"Failed to read ${lockFile}", err))
+    })
+    .flatMap { content =>
+      GradleLockFile.decodeGradleLockFile(new JawnParser, content) match {
+        case Right(m) => Success(m)
+        case Left(err) =>
+          Failure(new Exception(s"Failed to parse ${lockFile}", err))
       }
     }
 
   private def mergeLockFiles(
       lockFiles: List[GradleLockFile]
   ): Try[GradleLockFile] =
-    lockFiles.foldM(GradleLockFile.empty) { (s, n) =>
-      TryMerge.tryMerge(None, s, n)
+    lockFiles.foldM(GradleLockFile.empty) {
+      TryMerge.tryMerge(None, _, _)
     }
 
   // Gradle has compile/runtime/test sepearate classpaths
@@ -70,30 +73,38 @@ class GradleResolver(
       lockFile.testCompileClasspath,
       lockFile.testRuntimeClasspath
     )
-      .map(_.getOrElse(Map()))
-      .foldLeft(Try(Map[String, GradleLockDependency]())) { case (p, n) =>
-        p.flatMap { s => TryMerge.tryMerge(None, s, n) }
-      }
+    .map(_.getOrElse(Map.empty))
+    .foldM(Map.empty[String, GradleLockDependency]) {
+      TryMerge.tryMerge(None, _, _)
+    }
 
   private def assertConnectedDependencyMap(
       depMap: Map[String, GradleLockDependency]
   ): Try[Unit] =  {
-    val sunit = Success(())
+    // Use validated to collect all the missing errors, not just the first
+    type V[+A] = ValidatedNel[(String, String), A]
 
-    def assertDep(key: String, gld: GradleLockDependency): Try[Unit] =
-      gld.transitive.getOrElse(Nil).traverse_ { luK =>
+    def assertDep(key: String, gld: GradleLockDependency): ValidatedNel[(String, String), Unit] =
+      gld.transitive.getOrElse(Nil).traverse_[V, Unit] { luK =>
         if (!depMap.contains(luK)) {
-          Failure(new Exception(s"Unable to find $luK, referenced as a transitive dep for $key in dependencies."))
+          Validated.invalidNel((luK, key))
         }
         else {
-          sunit
+          Validated.Valid(())
         }
       }
 
-    depMap.toList.traverse_ { case (k, v) => assertDep(k, v) }
-  }
+    // sort to make this deterministically ordered
+    depMap.toList.sortBy(_._1).traverse_[V, Unit] { case (k, v) => assertDep(k, v) } match {
+      case Validated.Valid(()) => Success(())
+      case Validated.Invalid(errs) =>
+        Failure(new Exception("unable to find these referenced transitive deps in dependencies: " +
+          errs.toList.iterator.map { case (luK, key) => s"$key -> $luK" }.mkString("[", ", ", "]")
+        ))
+    }
+}
 
-  private def cleanUpMap(
+private def cleanUpMap(
       depMap: Map[String, GradleLockDependency]
   ): Map[String, GradleLockDependency] = {
     // for no content deps there is nothing to fetch/no sha to operate on.
@@ -102,7 +113,7 @@ class GradleResolver(
       val gradleProjectsRemoved = depMap.filter(_._2.project != Some(true))
 
       gradleProjectsRemoved
-        .foldLeft(Map[String, GradleLockDependency]()) { case (p, (nK, nV)) =>
+        .foldLeft(Map.empty[String, GradleLockDependency]) { case (p, (nK, nV)) =>
           @annotation.tailrec
           def go(parents: List[String], loop: Boolean, acc: List[String])
               : List[String] = {
@@ -171,13 +182,12 @@ class GradleResolver(
   //
   private def buildGraphFromDepMap(
     m: Model,
-      depMap: Map[String, GradleLockDependency]
+    depMap: Map[String, GradleLockDependency]
   ): Try[Graph[MavenCoordinate, Unit]] = {
     assertConnectedDependencyMap(depMap).map(_ => cleanUpMap(depMap)).flatMap { depMap =>
       def toCoord(k: String): Try[MavenCoordinate] =
-        depMap
-          .get(k)
-          .map { resolvedInfo =>
+        depMap.get(k) match {
+          case Some(resolvedInfo) =>
             val e = k.split(':')
             val (org, nme) = (e(0), e(1))
 
@@ -192,23 +202,19 @@ class GradleResolver(
                 Version(resolvedInfo.locked.getOrElse(""))
               )
             )
-          }
-          .getOrElse(
+          case None =>
             Failure(new Exception(s"Unable to lookup info about $k in dep map"))
-          )
+        }
 
-      val gradleDependencyGraphTry = depMap.foldLeft(Try(Graph.empty[MavenCoordinate, Unit])) {
-        case (tryG, (n, deps)) =>
+      val gradleDependencyGraphTry = depMap.toList.foldM(Graph.empty[MavenCoordinate, Unit]) {
+        case (g, (n, deps)) =>
           for {
-            g <- tryG
             cnode <- toCoord(n)
             transitive <- cats
               .Traverse[List]
               .sequence(deps.transitive.getOrElse(Nil).map(toCoord(_)))
           } yield {
-            val g1 = (cnode :: transitive).foldLeft(g) { case (p, n) =>
-              p.addNode(n)
-            }
+            val g1 = g.addAllNodes(cnode :: transitive)
             transitive.foldLeft(g1) { case (g, revDep) =>
               val curEdge = (
                 s"${revDep.group.asString}:${revDep.artifact.artifactId}",
@@ -225,31 +231,33 @@ class GradleResolver(
               }
             }
           }
-
-
       }
 
       gradleDependencyGraphTry.map { graph =>
-        m.dependencies.roots.foldLeft(graph) { case (g, n) =>
-          g.addNode(n)
-        }
+        graph.addAllNodes(m.dependencies.roots)
       }
     }
   }
+
+  def toTask[A](t: Try[A]): Task[A] =
+    t match {
+      case Success(value)     => Task.point(value)
+      case Failure(exception) => Task.fail(exception)
+    }
 
   // Build the entire transitive graph of a set of coordinates
   def buildGraph(
       coords: List[MavenCoordinate],
       m: Model
-  ): Task[Graph[MavenCoordinate, Unit]] = {
-    loadLockFiles(gradleTpe.getLockFiles)
-      .flatMap(mergeLockFiles(_))
-      .flatMap(collapseDepTypes(_))
-      .flatMap(buildGraphFromDepMap(m, _)) match {
-      case Success(value)     => Task.point(value)
-      case Failure(exception) => Task.fail(exception)
-    }
-  }
+  ): Task[Graph[MavenCoordinate, Unit]] =
+    toTask(
+      for {
+        lfs <- gradleTpe.getLockFiles.traverse(loadLockFile)
+        lockFile <- mergeLockFiles(lfs)
+        depMap <- collapseDepTypes(lockFile)
+        graph <- buildGraphFromDepMap(m, depMap)
+      } yield graph
+    )
 
   def resolve(model: Model): Task[
     (
@@ -258,18 +266,15 @@ class GradleResolver(
         Map[UnversionedCoordinate, Set[Edge[MavenCoordinate, Unit]]]
     )
   ] = {
-    buildGraph(Nil, model)
-      .flatMap { graph =>
-        def replaced(m: MavenCoordinate): Boolean =
-          model.getReplacements.get(m.unversioned).isDefined
-        for {
-          shas <- getShas(graph.nodes.filterNot(replaced).toList.sorted)
-        } yield (graph, shas, Map())
-      }
+    def replaced(m: MavenCoordinate): Boolean =
+      model.getReplacements.get(m.unversioned).isDefined
+
+    for {
+      graph <- buildGraph(Nil, model)
+      shas <- getShas(graph.nodes.filterNot(replaced).toList.sorted)
+    } yield (graph, shas, Map.empty)
   }
 
-  def run[A](fa: Task[A]): Try[A] = {
+  def run[A](fa: Task[A]): Try[A] =
     coursierResolver.run(fa)
-  }
-
 }
